tests/wpt/tests/css/css-fonts/parsing/font-style-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 4: parsing font-style with valid values</title>
<link rel="help" href="https://www.w3.org/TR/css-fonts-4/#font-style-prop">
<meta name="assert" content="font-style supports the full grammar 'normal | italic | oblique <angle [-90deg,90deg]>?'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value('font-style', 'normal');
test_valid_value('font-style', 'italic');
test_valid_value('font-style', 'oblique');

test_valid_value('font-style', 'oblique 10deg');
test_valid_value('font-style', 'oblique -10deg');
test_valid_value('font-style', 'oblique 0deg');
test_valid_value('font-style', 'oblique -90deg');
test_valid_value('font-style', 'oblique 90deg');
test_valid_value('font-style', 'oblique 10grad');
test_valid_value('font-style', 'oblique calc(90deg)');
test_valid_value('font-style', 'oblique calc(100deg)');
test_valid_value('font-style', 'oblique calc(-100deg)');
test_valid_value('font-style', 'oblique calc(30deg * 2)', 'oblique calc(60deg)');
test_valid_value('font-style', 'oblique calc(30deg + (sign(2cqw - 10px) * 5deg))', 'oblique calc(30deg + (5deg * sign(2cqw - 10px)))');
</script>
</body>
</html>
