third_party/blink/web_tests/external/wpt/css/css-fonts/parsing/font-style-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['font-style'] = "oblique calc(100deg)" should set the property value
  assert_equals: serialization should be canonical expected "oblique calc(100deg)" but got "oblique 90deg"
[FAIL] e.style['font-style'] = "oblique calc(-100deg)" should set the property value
  assert_equals: serialization should be canonical expected "oblique calc(-100deg)" but got "oblique -90deg"
Harness: the test ran to completion.
